server/requirements.txt
Flask==2.2.3
Flask-RESTful==0.3.10
gunicorn==20.1.0
firebase-admin==6.1.0
python-dotenv==1.0.0